requirements.txt
fastapi
python-multipart
uvicorn
openai
python-dotenv
qdrant-client